modificando el redireccionamiento a whatsapp (logica de firebase en los viewmodels)
