+++ Servitodo/.idea/misc.xml
@@ -35,7 +35,7 @@
         <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/activity_cliente.xml" value="0.2" />
         <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/activity_main.xml" value="0.2" />
         <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/fragment_detalle_publicacion.xml" value="0.33" />
-        <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/fragment_home_cliente.xml" value="0.25" />
+        <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/fragment_home_cliente.xml" value="0.33" />
         <entry key="..\:/SERVITODO/Servitodo/app/src/main/res/layout/fragment_publicacion.xml" value="0.2" />
       </map>
     </option>
