+++ Servitodo/app/src/main/java/com/ort/servitodo/fragments/login/LogInFragment.kt
@@ -10,9 +10,10 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.navigation.findNavController
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
 import com.ort.servitodo.R
+
+//import com.google.firebase.firestore.ktx.firestore
+//import com.google.firebase.ktx.Firebase
 import com.ort.servitodo.entities.Usuario
 import com.ort.servitodo.viewmodels.login.LogInViewModel
 
@@ -41,28 +42,6 @@ class LogInFragment : Fragment() {
         super.onStart()
 
 
-        val db = Firebase.firestore
-
-        // Create a new user with a first and last name
-        /*val user = Usuario(
-            id = 1,
-            nombre = "Lovelace",
-            apellido = "born"
-        )*/
-
-        //db.collection("cities").document("new-city-id").set(user)
-
-
-        // Add a new document with a generated ID
-        /*db.collection("users")
-            .add(user)
-            .addOnSuccessListener { documentReference ->
-            Log.e("Buenas", documentReference.toString())
-            //Log.d(TAG, "DocumentSnapshot added with ID: ${documentReference.id}")
-            }
-            .addOnFailureListener { e ->
-                Log.w(TAG, "Error adding document", e)
-            }*/
 
         btnLogin.setOnClickListener {
             val action = LogInFragmentDirections.actionLogInFragmentToPrestadorActivity()
