+++ Servitodo/app/src/main/java/com/ort/servitodo/viewmodels/cliente/DetallePublicacionViewModel.kt
@@ -2,9 +2,13 @@ package com.ort.servitodo.viewmodels.cliente
 
 import android.content.Intent
 import android.content.pm.PackageManager
+import android.net.Uri
 import android.view.View
 import androidx.lifecycle.ViewModel
+import androidx.navigation.findNavController
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.snackbar.Snackbar
+import com.ort.servitodo.R
 import com.ort.servitodo.entities.Prestador
 import com.ort.servitodo.entities.Publicacion
 import com.ort.servitodo.repositories.PrestadorRepository
@@ -36,21 +40,36 @@ class DetallePublicacionViewModel : ViewModel() {
         val msg = "Hola. Te quiero contratar"
         val packageWhatsApp = "com.whatsapp"
 
-        if(isAppInstalled(packageWhatsApp)){
+        //if(isAppInstalled(packageWhatsApp)){
 
+            /*--> TODO: Opcion 1
             val sendIntent = Intent().apply {
                 action = Intent.ACTION_SEND
                 putExtra(Intent.EXTRA_TEXT, msg)
                 putExtra("jid", "${prestador.numtel}@s.whatsapp.net")
+                //putExtra("jid", "https://api.whatsapp.com/send/?phone=${prestador.numtel}")
                 type = "text/plain"
                 setPackage(packageWhatsApp)
             }
+            */
+
+            /*--> TODO: Opcion 2
+            val gmnIntentUri = Uri.parse("https://api.whatsapp.com/send?phone=${prestador.numtel}&text=${msg}")
+            val sendIntent = Intent(Intent.ACTION_VIEW, gmnIntentUri)
+            sendIntent.setPackage(packageWhatsApp)
+            view.context.startActivity(sendIntent)
+            */
+
+            //--> TODO: Opcion 3
+            val gmnIntentUri = Uri.parse("https://wa.me/${prestador.numtel}?text=${msg}")
+            val sendIntent = Intent(Intent.ACTION_VIEW, gmnIntentUri)
+            sendIntent.setPackage(packageWhatsApp)
             view.context.startActivity(sendIntent)
 
-        }
+        /*}
         else{
             Snackbar.make(view, "Debes instalar Whatsapp", Snackbar.LENGTH_SHORT).show()
-        }
+        }*/
     }
 
     //--> Chequea que la App este instalada
@@ -62,5 +81,19 @@ class DetallePublicacionViewModel : ViewModel() {
         } catch (e: PackageManager.NameNotFoundException) {
             false
         }
+    }
+
+    //--> Pop up para confirmar el redireccionamiento 
+    fun confirmRedirectionToWhatsapp(index : Int){
+        MaterialAlertDialogBuilder(view.context)
+            .setTitle("Confirmar")
+            .setMessage("Deseas ser redireccionado a whatsapp?")
+            .setNegativeButton("Cancelar") { dialog, which ->
+                //view.findNavController().navigateUp()
+            }
+            .setPositiveButton("Aceptar") { dialog, which ->
+                this.whatsapp(index)
+            }
+            .show()
     }
 }
