+++ Servitodo/app/src/main/java/com/ort/servitodo/viewmodels/login/LogInViewModel.kt
@@ -3,5 +3,28 @@ package com.ort.servitodo.viewmodels.login
 import androidx.lifecycle.ViewModel
 
 class LogInViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
+
+    //val db = Firebase.firestore
+
+    // Create a new user with a first and last name
+    /*val user = Usuario(
+        id = 1,
+        nombre = "Lovelace",
+        apellido = "born"
+    )*/
+
+    //db.collection("cities").document("new-city-id").set(user)
+
+
+    // Add a new document with a generated ID
+    /*db.collection("users")
+        .add(user)
+        .addOnSuccessListener { documentReference ->
+        Log.e("Buenas", documentReference.toString())
+        //Log.d(TAG, "DocumentSnapshot added with ID: ${documentReference.id}")
+        }
+        .addOnFailureListener { e ->
+            Log.w(TAG, "Error adding document", e)
+        }*/
+
 }
