+++ Servitodo/app/src/main/java/com/ort/servitodo/viewmodels/login/SignUpViewModel.kt
@@ -1,7 +1,41 @@
 package com.ort.servitodo.viewmodels.login
 
 import androidx.lifecycle.ViewModel
+//import com.google.firebase.firestore.ktx.firestore
+//import com.google.firebase.ktx.Firebase
 
 class SignUpViewModel : ViewModel() {
     // TODO: Implement the ViewModel
+
+    /*   val db = Firebase.firestore
+
+    // Create a new user with a first and last name
+    val user = hashMapOf(
+        "first" to "Ada",
+        "last" to "Lovelace",
+        "born" to 1815
+    )
+
+
+    // Add a new document with a generated ID
+    val s = db.collection("users")
+    .add(user)
+    .addOnSuccessListener { documentReference ->
+        //Log.d(TAG, "DocumentSnapshot added with ID: ${documentReference.id}")
+    }
+    .addOnFailureListener { e ->
+        //Log.w(TAG, "Error adding document", e)
+    }*/
+
+    //Get
+    /*val asda = db.collection("users")
+    .get()
+    .addOnSuccessListener { result ->
+        for (document in result) {
+            //Log.d(TAG, "${document.id} => ${document.data}")
+        }
+    }
+    .addOnFailureListener { exception ->
+        //Log.w(TAG, "Error getting documents.", exception)
+    }*/
 }
